apply comments by Copilot
