+++ test/tateyama/endpoint/ipc/crypto/rsa.h
@@ -13,6 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#pragma once
+
 #include <memory>
 #include <string>
 #include <string_view>
